requirements.txt
torch
torchvision
numpy
opencv-python
nibabel
fastapi
uvicorn
streamlit
requests
matplotlib
scikit-image
pillow
